split up files and organize code
