+++ frontend/src/scenes/events/EventsTable.tsx
@@ -21,7 +21,7 @@ import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { EventFormattedType } from '~/types'
 import { PageHeader } from 'lib/components/PageHeader'
 import { TableConfig } from 'lib/components/ResizableTable'
-import { propertyDefinitionsLogic } from './propertyDefinitionsLogic'
+import { propertyDefinitionsLogic } from './volume-definitions/propertyDefinitionsLogic'
 import { EventName } from 'scenes/actions/EventName'
 import { PropertyFilters } from 'lib/components/PropertyFilters'
 import { FEATURE_FLAGS } from 'lib/constants'
