+++ frontend/src/scenes/events/volume-definitions/VolumeTable.scss
